Move more simplify/detail methods to their own classes
