+++ syft/serde/serde.py
@@ -80,14 +80,8 @@ from syft.serde.torch_serde import (
     _detail_dictionary,
     _detail_ndarray,
     _detail_torch_device,
-    _detail_pointer_tensor,
-    _detail_log_tensor,
-    _detail_additive_shared_tensor,
-    _detail_multi_pointer_tensor,
-    # _detail_plan,
     _detail_worker,
     _force_full_detail_worker,
-    _detail_object_wrapper,
     _detail_script_module,
     _detail_exception,
     _detail_train_config,
@@ -97,14 +91,8 @@ from syft.serde.torch_serde import (
     _simplify_dictionary,
     _simplify_ndarray,
     _simplify_torch_device,
-    _simplify_pointer_tensor,
-    _simplify_log_tensor,
-    _simplify_additive_shared_tensor,
-    _simplify_multi_pointer_tensor,
-    # _simplify_plan,
     _simplify_worker,
     _simplify_exception,
-    _simplify_object_wrapper,
     _force_full_simplify_worker,
     _simplify_script_module,
     _simplify_train_config,
@@ -409,14 +397,14 @@ simplifiers = {
     slice: [8, _simplify_slice],
     type(Ellipsis): [9, _simplify_ellipsis],
     torch.device: [10, _simplify_torch_device],
-    pointers.PointerTensor: [11, _simplify_pointer_tensor],
-    LoggingTensor: [12, _simplify_log_tensor],
-    AdditiveSharingTensor: [13, _simplify_additive_shared_tensor],
-    MultiPointerTensor: [14, _simplify_multi_pointer_tensor],
+    pointers.PointerTensor: [11, sy.PointerTensor._simplify_pointer_tensor],
+    LoggingTensor: [12, sy.LoggingTensor._simplify_log_tensor],
+    AdditiveSharingTensor: [13, sy.AdditiveSharingTensor._simplify_additive_shared_tensor],
+    MultiPointerTensor: [14, sy.MultiPointerTensor._simplify_multi_pointer_tensor],
     Plan: [15, sy.Plan._simplify_plan],
     VirtualWorker: [16, _simplify_worker],
     str: [18, _simplify_str],
-    pointers.ObjectWrapper: [19, _simplify_object_wrapper],
+    pointers.ObjectWrapper: [19, sy.ObjectWrapper._simplify_object_wrapper],
     GetNotPermittedError: [20, _simplify_exception],
     ResponseSignatureError: [20, _simplify_exception],
     torch.jit.ScriptModule: [21, _simplify_script_module],
@@ -468,15 +456,15 @@ detailers = [
     _detail_slice,
     _detail_ellipsis,
     _detail_torch_device,
-    _detail_pointer_tensor,
-    _detail_log_tensor,
-    _detail_additive_shared_tensor,
-    _detail_multi_pointer_tensor,
+    sy.PointerTensor._detail_pointer_tensor,
+    sy.LoggingTensor._detail_log_tensor,
+    sy.AdditiveSharingTensor._detail_additive_shared_tensor,
+    sy.MultiPointerTensor._detail_multi_pointer_tensor,
     sy.Plan._detail_plan,
     _detail_worker,
     _force_full_detail_worker,
     _detail_str,
-    _detail_object_wrapper,
+    sy.ObjectWrapper._detail_object_wrapper,
     _detail_exception,
     _detail_script_module,
     _detail_train_config,
